requirements.txt
matplotlib
numpy
pillow
pygame
pyk4a
pykalman
pyyaml
pandas
open3d
scipy
statsmodels==0.13.2
requests